Rename classes for better readability
