+++ testar/src/org/testar/reporting/PlainTextFormatUtil.java
@@ -30,95 +30,70 @@
 
 package org.testar.reporting;
 
+import org.apache.commons.lang.StringUtils;
 import org.testar.monkey.Assert;
 
 import java.util.Collections;
 import java.util.List;
 
-public class HTMLreportUtil extends BaseReportUtil
+public class PlainTextFormatUtil extends BaseFormatUtil
 {
-    public HTMLreportUtil(String filePath)
+
+    protected PlainTextFormatUtil(String filePath)
     {
-        super(filePath, "html");
+        super(filePath, "txt");
     }
     
     public void addContent(String text)
     {
-        if(text.contains("\n"))     Collections.addAll(content, splitStringAtNewline(text));
-        else                        content.add(text);
-    }
-    
-    public void addContent(String text, String tag)
-    {
         
         if(text.contains("\n"))
-        {
-            content.add("<" + tag + ">");
             Collections.addAll(content, splitStringAtNewline(text));
-            content.add("</" + tag + ">");
-        }
         else
-            content.add("<"+tag+">"+text+"</"+tag+">");
+            content.add(text);
     }
-    
-    public void addHeader(String title)
-    {
-        addHeader(title, "");
-    }
-    
-    public void addHeader(String title, String script)
-    {
-        Assert.notNull(script);
-        Assert.notNull(title);
-    
-        content.add("<!DOCTYPE html>");
-        content.add("<html lang=\"en\">");
-        content.add("<head>");
-        if(!script.equals(""))
-            addContent(script, "script");
-        content.add("<meta http-equiv=\"Content-Type\" content=\"text/html; charset=utf-8\">");
-        addContent(title, "title");
-        content.add("</head>");
-        content.add("<body>");
-    }
-    
-    public void addFooter()
-    {
-        content.add("</body>");
-        content.add("</html>");
-    }
-    
     
     public void addHeading(int level, String text)
     {
-        Assert.isTrue(level >= 1 && level <= 6, "Invalid HTML heading level");
+        Assert.isTrue(level >= 1 && level <= 6, "Invalid heading level: must be between 1 and 6");
         Assert.notNull(text);
-        addContent(text, "h" + level);
+        content.add(StringUtils.repeat("#", level) + text); //repeat # char
     }
-    
-    public void addParagraph(String text) {
-        Assert.notNull(text);
-        addContent(text, "p");
-    }
-    
-    public void addLineBreak()
+
+    public void addParagraph(String text)
     {
-        content.add("<br>");
+        Assert.notNull(text);
+        addEmptyLine();
+        Collections.addAll(content, splitStringAtNewline(text));
+        addEmptyLine();
+    }
+
+    public void addEmptyLine()
+    {
+        content.add("");
+    }
+
+    public void addHorizontalLine()
+    {
+        addHorizontalLine(20);
+    }
+
+    public void addHorizontalLine(int length)
+    {
+        Assert.isTrue(length >= 3 && length <= 100, "Invalid horizontal line length: must be between 3 and 100");
+        content.add(StringUtils.repeat("=", length)); //repeat # char
     }
     
     public void addList(boolean ordered, List<String> items)
     {
-        if(ordered)
-            content.add("<ol>");
-        else //unordered
-            content.add("<ul>");
-    
-        for(String item : items)
-            addContent(item,"li");
-        
-        if(ordered)
-            content.add("</ol>");
-        else //unordered
-            content.add("</ul>");
+        addEmptyLine();
+        for(int i = 0; i < items.size(); i++)
+        {
+            if(ordered)
+                addContent(String.valueOf(i+1) + items.get(i));
+            else
+                addContent("- " + items.get(i));
+        }
+        addEmptyLine();
     }
 }
